+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Avoid linebreaks between a heading and a captioned element.
+
 ## [1.2.1] 2019-09-19
 
 -   Keep caption on same page as figure